Set time zone for PhabricatorRepositoryCommitDiscoveryDaemon
Summary:
the daemon is generating a lot of warning about timezone not
being set. Add it.

Test Plan:
php -l. Launched the daemon and it didn't crash. It's
difficult to make it parse a commit on my sandbox though. I will see if
it fixes the problem once it is live.

Reviewed By: epriestley
Reviewers: epriestley, tuomaspelkonen
CC: aran, epriestley, jungejason
Differential Revision: 356