Fix regenerate arcanist cert and setup stuff
Summary: well, since I couldn't regenerate my arcanist cert I figured out that
this was because "workflows" are unavailable there now. I really can not figure
out why but it was.
By the other hand, I added the ability to add a trailing slash at the end of
the hostname when it isn't present there while we are generating the cert, and
also I added in the setup script, the ability to check if is present the
protocol of the host since it is needed to generate the cert as well as the
trailing slash.

Well those things were fired up, because I forgot to set the protocol and the
slash on the config file, since someone could make the same mistakes. so I
tried to minimize them.

Test Plan:

FIRST: DO NOT apply those changes! then
1- go to settings->arcanist certificate and the click on regenerate ... humm
2- On your config file, delete the trailing slash at the end and the protocol
on "phabricator.base-uri", then go to setting->arcanist certificate. Here you
will see something like this "phabricator.example.comapi\/" instead of
"http:\/\/phabricator.example.com\/api\/".

SECOND: Now apply this changes:
1- Go to settings->arcanist certificate and the click on regenerate.
2- On your config file, delete the trailing slash at the end and the protocol
on "phabricator.base-uri", and setup "phabricator.setup" to true.
3- Do not add the trailing slash. Then go to setting->arcanist certificate and
you could see that we added a trailing slash before "api".

Reviewers: epriestley jungejason

CC: epriestley jungejason