Allow standard page body panel to scroll on overflow-x
Summary:
This seems like the least-bad solution to the issues mentioned in T684: when we
need to x-scroll the main page area, scroll that div rather than the surrounding
page chrome.

I played around with a bunch of other possible solutions but they all seem bad
in some way or another. The tricky part here is that I want the real background
to be grey so that the footer color is grey even if the page is very short and
the browser window is very tall.

The only downside here is that the scrollbar appears in a somewhat unusual
place, but I think that's OK?

Actually, it's kind of terrible if people really use the scrollbar to scroll
horizontally rather than two-finger swipe or shift+mousewheel or the arrow keys.
So maybe this isn't good.

If this is no good, I think we need to make design sacrifices (not necessarily a
big deal; I'm not married to how the footer behaves) or someone much better than
I am at CSS needs to tell me how to fix this (@mroch / @tomo)?.

Test Plan:
  - In Settings -> Preferences, set font to "72px Impact".
  - Observed overflow scroll behavior in Safari / Firefox / Chrome.

Reviewers: Makinde, btrahan, jungejason

Reviewed By: Makinde

CC: mroch, tomo, aran, Makinde

Maniphest Tasks: T684

Differential Revision: 1224