Move config files out of custom/.
I understand now why the instructions had you put these files in
custom -- custom is in .gitignore, so you could have these files and
not worry about git getting confused.

But I *want* these files to be in git; we have our own fork of the
phabricator source.  So I don't want them to be ignored, so there's no
real reason to put them in custom.