Removing memory limit sanity check
Summary:
Turns out that 12 characters is not enough for folks that have
memory_limit set to INT64_MAX (9223372036854775807).

Since this only seems to have affected a single installation,
epriestley says let's get rid of it.  If it comes up again, we can
restore the check but use a bigger number.

Test Plan: Just make a call to the phabricator webroot.

Reviewers: epriestley

Reviewed By: epriestley

CC: aran, Korvin

Maniphest Tasks: T1641

Differential Revision: https://secure.phabricator.com/D3246